payment-consumer/requirements.txt
Flask==2.3.1
redis==4.5.4
kafka-python